feat: Add core features to client and asyncio server
Update client.py so it no longer acts as a pipe that just sends user
input to the asyncio server. client.py now cycles through client.csv on
user input and sends a bytestream and lboxes array for each image. This
update implements the core functionality we want from the client in that
it sends over an image bytestream and metadata, similar to what scrubcam
does.

The asyncio server code was also updated in app_V2.py to read in the
image bytestream and lboxes array messages. The asyncio server writes
the image to disk and adds the entry to image_log.csv. It then packages
the saved image filename and image class name in a dictionary and adds
it to the shared queue that sends the information to the dash server.