feat: Dash server initialize grid images from csv
The dash server now creates the initial image dictionary by setting the
value for each animal to the most recent image recorded in
image_log.csv. If no image exists for the animal, the value will be
'None' and the 'no image' icon is rendered on the webpage.

image_log.csv has 24 entries (2 for each animal). The saved_images/
directory contains the 24 files recorded in the csv. This allows the
dash server to initialize an image for each animal and also shows that
the image is initialized with the most recent image.

Finding the most recent image can be done by sorting the image path in
descending order since the timestamp is the first part of the filename.
This means that the most recent file is listed first.