feat: Implement history page for each animal
The dash server now supports a multi-page app that takes the user to a
'history page' after clicking on an animal's image in the main grid.

The history page contains the animal to be shown in the url pathname.
The history page uses the url to determine which animal to show and
reads image_log.csv to render all the images for the animal in most
recent to least recent order.