feat: Update client to send class list on startup
** Client **

client.py now sends the classification list on startup before sending
images over the socket. This was implemented by sending a header first.
The header can take the value "CLASSES" or "IMAGE" and describes what
the message contents will be. Then the appropriate contents are sent.

** Asyncio Server **

The asyncio server was updated so it reads in the header first and based
on the value, it calls either the handle_image() or handle_classes()
method to correctly read the remainder of the message.

Before, the asyncio server would communicate to the dash server by
sending a dictionary that looked like: {"img_path": 'some_path.png',
"class_name": 'some_classification'}. This was okay since only images
were being shared.

Now, the header is included as a key so dash knows whether the content
it reads from the shared Queue is an image or a class list. A message
for the class list is still communicated as a dictionary to keep the
representation consistent.

** Dash Server **

The dash server now shows a "Waiting to connect" message on startup
until it receives the classification list from the client. The grid will
only be shown after the classification list has been received.

To persistently store the classification list between the history page
and the grid page, I had to add the storage_type='local' argument to the
'image-dict' dcc.Store component. This allows the dictionary to be kept
after the browser quits. However, clearing cookies will clear the
dictionary.

Note: The classification list for the grid can change dynamically if the
client sends over a new classification list.