refactor: Split app_V2.py into 3 separate files
app_V2.py was split into three separate files:
  1) scrubdash.py
  2) asyncio_server.py
  3) dash_server.py

** Scrubdash **

scrubdash.py only contains the entrypoint into the program. It only
consists of a main method and will create the asyncio server and dash
server via multiprocessing.

** Asyncio Server **

asyncio_server.py contains asyncio server code. Before, in app_V2.py,
the asyncio server code was implemented as a function. However, upon
splitting the code, I found it better to refactor the asyncio server
function to be a class that takes in the shared queue as an instance
parameter. This provides better encapsulation.

** Dash Server **

dash_server.py contains dash server code. The code is a direct
copy-paste from app_V2.py. I tried to refactor the start_dash() function
into a class but that did not work because the callback function
decorators need to start with @app. However, this would not be possible
if the dash server was implemented as a class since it would become
@self.app, which would not be recognized by the dash library.

I also tried to structure the dash server using multiple files. However,
I also ran into difficulties since the documentation recommends index.py
to be the entrypoint into the app with __name__ == '__main__'. This
would not allow me to pass the shared queue as a parameter to the dash
server since the main method does not take parameters.

Splitting the dash server code into separate files is most likely
possible - I just need to find the correct way to partition the code.