feat: Implement data visualization page
The dash server main page supports a data visualization page, accessible
via a link on the grid page.

** Design Choice **

To share the classes between the grid page and visualization page, the
image dictionary is now stored in app.layout instead of index_page.

I found this less convoluted than passing the image dictionary as a
state parameter in the display_page() callback and adding an image-dict
parameter to the create_x_page() functions. If implementing the latter,
I have to modify the index_page variable to become a function so it
conforms to the other create_x_page() functions, which seems unnecessary
at the moment.

** Bugs / Weird Behaviors **

The time histogram has some behaviors that should be noted:
  - if no images were received during the time span specified, the graph
    can show undefined behavior, such as the x-axis format getting
    messed up.
  - if the start date specified from the time span dropdown is before
    the date of the first image received, the graph starts at the date
    of the image, not the start date argument given.