feat: Draw rectangles on dash history images
This is not an atomic commit. This commit includes a new feature and
refactored code.

The new feature is the dash history page now draws rectangles around
objects in each image that are above the confidence threshold specified
by a slidable toggle. The class name of the object as well as the
classification confidence is rendered as text within each rectangle.

dash_server.py is refactored to be a multi-page program to better
encapsulate code based on functionality.