feat: Optimize history page rendering
*** Layout ***

Images are now displayed in pages of 3x3 grids. One page is shown at a
time and users can cycle through pages with the `Next` and `Back`
buttons.

In previous versions, the image grid was one monolithic component. All
images had to finish loading before anything was displayed. In the 3x3
grid, each image is its own component. This allows images to render once
they are done loading and they do not need to wait for other images to
load before being displayed.

*** Modal Component ***

Clicking on an image opens a modal component that displays the
confidence slider and a larger version of the image just clicked on.
Users can change the value on the slider to change the threshold used to
draw lboxes onto the image.

In previous versions, lboxes were being drawn on every image in the
grid. By focusing on only a single image now, less computation and time
is required to draw lboxes on the image.