feat: Implement asyncio continue run flag
Asyncio finds the most recently modified "session folder" within
RECORD_FOLDER and saves images and lboxes to that "session folder."
Asyncio appends image metadata to the already existing image log for
that session as well.

Functionality is finished for both a new run and continuing run. The
path for the session's image log is not being sent over to dash yet, but
everything is implemented on the asyncio side.