feat: Implement new run flag for dash server
The dash server now has functionality to service a new run and a
continuing run.

The image log path and the class list need to be sent from the asyncio
server to the dash server. These variables should act as global
variables in dash, which allows the website to be closed and reopened
without losing any state.

Storing these variables in a dcc.Store() component is not persistent
unless a `storage_type` property is specified. However, all storage_type
options store the data on the client side, which we want to avoid.
Having the image log path and class list act as global variables mimics
storing the data on the server side, which is the ideal behavior we
want.

Since we want these variables known on dash server start-up, and since
these values are initialized by the asyncio server, we need a way for
scrubdash to obtain the image log path and class list from asyncio and
pass these to dash. I achieved this by having scrubdash pass these
values as parameters to dash_server.start_dash().

I considered passing the image log path and class list from asyncio to
dash directly through the shared queue, but this would not allow dash to
access these variables on start-up; they would only be accessed by the
callback that checks the shared queue. While the shared queue
implementation may work, we would have to ensure that asyncio passes
these values to dash in a timely manner, or else dash may initially
render in an incorrect state. Having the image log path and class list
go through scrubdash first ensures that asyncio is finished configuring
all the user session folders and files before starting the dash server,
which is less error prone.

The log_queue variable is used in scrubdash.py so scrubdash can receive
the image log path and filter class values from asyncio. I tried to
implemented a getter method for each of these values, but the incorrect
value was being returned since all these values are initialized after
starting the asyncio server, which runs in an entirely different
process. Therefore, the correctly initialized variables are not
accessible through a getter method.