feat: Implement detected classes in asyncio and dash
The 'label' column in the image log is changed to 'labels'. Instead of
saving only the class with the highest confidence in the image, the
asyncio server now saves a list of all unique classes observed in the
image in the column 'labels'.

This change allows images that feature more than one class in
filter_classes to render in multiple history pages and to appear as the
thumbnail picture for multiple classes on the grid page.

This change breaks the graphs page, as values in the 'label' column were
used as values for the x-axis. This worked when the 'label' value was
only a single class. However, the 'labels' value is now a list that
usually consists of several classes. This is not desirable since it now
aggregates values based on a set representation of the classes in an
image. The only way to create the same graphs as before would be to
transform the current image log to a representation that counts the
number of instances of each class in the image log.

This brings up another issue since asyncio only lists the unique classes
in an image. This undercounts classes that appear multiple times in an
image. However, several of these classifications can be false positives
(eg. a person is classified when there is not a person in the lbox).
Further discussion on this topic is necessary before fixing the graphs
page.