feat: Rename cont. flag and update graph page
*** New Flag ***

The --new flag is replaced by the --cont flag. This is changed because
the assumption is that a user will create a new session by default. To
continue the most recent session, the --cont flag must be added.

*** Graph Page Architecture Design Choice ***

The graph page is updated to be compatible with the new format of the
image log. Compatability requires access to the filter_classes list. Two
viable options are to pass the list through the shared queue between
asyncio and dash or to pass the list first from asyncio to scrubdash
through the shared log_queue queue, and then from scrubdash to dash
through a parameter to dash_server.start_dash().

The first option is more favorable since the second option will block
scrubdash if the scrubcam is not connected since it will be waiting for
the value from asyncio via the log_queue.

The first option does not block if scrubcam is not connected yet.
However, the filter-class 'data' component will be updated every 1.5
seconds when the shared queue between asyncio and dash is checked. This
has no noticeable affect on the graphs page as only the dropdown option
callback is triggered and once the filter class list has been sent, the
value should remain unchanged.

*** Plotly Graph Compatability Implementation ***

The plotly histogram constructor does not support the new format of the
image log with a labels list, so I transformed the image log into a new
dataframe that flattens the list in the 'labels' column of the image log
such that each row contains only one label and the datetime it was
recorded. This allows the histogram to count the number of rows a label
appears in and to appropriately plot each instance based on its
datetime. This transformed dataframe is created every time the graph
page is loaded and is stored in the 'label-count' dcc.Store component.

The time series histogram x-axis labeling is also fixed. Before, it
would create a tick for every interval specified by `selected_interval`.
This would be unreadable if given a long `selected_span` (eg. 1 month)
and a short `selected_interval` (eg. 30 minutes). Changing the parameter
to fig.update_xaxes(tickmode='auto') and not specifying `nticks` seems
to automatically create ideally spaced ticks that is readable.