feat: Implement no results found message
The graph page is updated to show the text "No matching data found" on a
blank white background as the figure for the time-graph component when
no images are found for the selected span.

This fixes the previous problem when no images are found for the
selected span, where the figure's x-axis start bin would be some
nonsensical value, such as a different year (eg. 2003). No data would be
displayed, but the x-axis labelings were all incorrect when no data was
found.

This fix allows the figure of the time-graph component to be used when
images are found and not found. The contents to the figure are the only
parameters that change and no further changes are necessary.