feat: Implement asyncio server email notification
The asyncio server now sends an email to a recipent when a class in
ALERT_CLASSES is detected in an image received from scrubcam. The email
alerts the recipient which classes from ALERT_CLASSES are detected and
also contains an attached of the image.

ALERT_CLASSES is specified in the config file. Several other keys were
added to the config file that specify who the sender and recipient of
the email notifications are.