feat: Update file structure to create pypi package
An asyncio_server folder was created to store the code for the asyncio
server and for email notifications. This helps consolidate the asyncio
code better.

The multi_page_dash_server folder was renamed to dash_server for
readability.

A symbolic link named __main__.py was created as an entry-point for
scrubdash to be run as a script. __main__.py points to scrubdash.py in
the outer folder. This mirrors the file structure of scrubcam.

testpypi.pw contains the password to publish the scrubdash package to
testpypi.