feat: Implement SMS notifications
The asyncio server sends an SMS notification every time a received image
contains classes listed in ALERT_CLASSES.

The example config file is updated to include an SMS receiver key-value
pair. The value is a list of dictionaries that represent all the people
who should receive an SMS notification. Each dictionary includes a
receiver's phone number and service carrier. An SMS notification is sent
to each receiver in this list.

The aiosmtplib library is chosen to send SMS notifications since the
smtplib library implementation is not working. The smtplib
implementation sends the SMS from gmail, but my phone never receives the
message. The gmail sent box records sending the SMS notification though,
so I'm not sure what the problem is.