feat: Add notification cooldown period (fix #11)
The previous classes `sms_sender` and `email_sender` are aggregated into
a single class `notification`. This is due to both `sender` classes
having very similar attributes. There is not much added benefit in
creating two separate classes when their functionalities could be
implemented as different methods in a more general `notification` class.

To prevent notification spam when several images are taken in a short
interval, the asyncio server only sends an image every 60 seconds. This
cooldown duration is taken from the `COOLDOWN_TIME` key in config file,
a key that was just added in this commit.

The cooldown time is not sent as an initialization parameter to
notification and is instead handled by the asyncio server. This is to
improve referential integrity. Whenever `send_email()` and `send_sms()`
are called by `notification`, we expect a notification to be sent. If
`notification` handles the cooldown, some calls to `send_email()` and
`send_sms()` would send a notification while other calls would not and
it would be confusing to someone new to the codebase.

Therefore, the asyncio server checks if the cooldown period has passed
before calling the notification methods. This also mirrors the logic in
`scrubcam`, where scrubcam only calls the `socket_handler` to send an
image if an observed class is in `FILTER_CLASSES`. The outer program
handles the business logic before calling a more specialized method.