feat: Add color picker and confidence range slider
A collapsable color-picker is available in the modal component that
allows users to change the font color of the lboxes drawn on the image
to improve readability on all background colors.

The confidence slider is now a range slider. Instead of only having a
maximum value, the range slider has a minimum and maximum value. Only
lboxes with a confidence between the min and max range will be drawn
onto the image.