refactor: Standardize parameters from config file
*** Config File Parameters ***

Previously, a module would first read specific values from the config
file before invoking a method or initializing a class with parameters
obtained from the config file.

When this pattern occurs in ScrubCam, the config file is read into a
dictionary in the invoker module, and the whole dictionary is passed as
a single parameter to the invokee class or method. This allows the
class/method to get the values it needs by itself and declutters the
paramter signature.

To make code consistent between ScrubCam and ScrubDash, a single config
dictionary is passed as a parameter to a class/method that needs values
from the config file. The class/method can index into the config
dictionary to obtain any values it needs and reduces the number of
parameters needed for some classes and methods.

** Asyncio Server Refactoring ***
The `asyncio_server` class is renamed to `AsyncioServer` to comply with
PEP8 class naming convention.

Docstrings are updated and more comments are added to code that may seem
complicated.