refactor: Standardize class naming convention
Class names are standardized in the asyncio_server package. The
`notification` class is renamed to `NotificationSender` and the
`session` class is renamed to `HostSession`. Classes are renamed to be
more descriptive and to conform to PEP8 CamelCasing convention for
classes.

Imports are now sorted in lexicographical order using the `isort`
Python utility. Docstrings are also updated for all modules in the
asyncio_server package.