feat: Implement pagination for the history page
Users can now see the total number of 3x3 grids available to sift
through on the history page. Instead of just relying on the `Next` and
`Back` buttons to flip through the different grids, users can now input
the grid number they want to see.

For example, if a user wanted to go to page 5, they can just input 5
instead of clicking the `Next` button four times. An error message also
appears if the input value is less than 1 or greater than the total
number of available grids.