feat: Change confidence slider values + add labels
The confidence slider values are now [0, 100] instead of [0, 1] to
represent percentage values instead of probability values. This makes
the slider more human readable.

In addition, tick marks label every 5% interval, allowing users to
better navigate the slider. The min and max value also render as a
tooltip so users can see the selected value as they move the slider.

Lastly, the confidence percentage is truncated to three decimal places
when rendered as text annotation for the drawn lboxes. This decreases
the clutter drawn on the image.