chore: Create symbolic link in package to configs
A symbolic link is made in the scrubdash package that links to the top
level config file. This is done so the package has a handle to the
config file after being built as a distribution.

A symbolic link pointing to a higher level file is chosen over creating
a link to an inner file since linking to an outer file is the
precedent (scrubdash/__main__.py links to the outer scrubdash.py file).