style: Style the home/hosts page
The hosts page (which will be renamed to home page) is now styled to
mirror the SDZWA website design. The page is also responsive to
different screen widths.

A navbar that will be used for all pages is in navbar.py. There are
three links: 'HOME', 'LABELS', and 'GRAPHS'. Only the 'HOME' link is
set to be visible by default. The logic is that callbacks for the
labels, graphs, and history page will make the 'LABELS' and 'GRAPHS'
links visible once the user has chosen a ScrubCam since the appropriate
device-specific href can now be created. Making the latter two links
invsible by default allows the same navbar code to be used for all
pages, even though the links that are accessible will differ depending
on the page.