style: Style the grid/labels page
The grid page is styled in the same fashion as the main page.

The navbar idea from the previous commit will not work. There are too
many wrinkles involved when trying to unhide the 'LABELS' and 'GRAPHS'
links. Therefore, I made two different navbars: a default one that just
has the 'HOME' link and a full navbar that has the 'HOME', 'LABELS',
and 'GRAPHS', link. This solution is easier to implement than trying to
make the {'display': 'none'} idea work.