style: Add credits page and GitHub page
A credits page is created to give credits to LogoMakr.com for creating
the ScrubDash logo.

A link to the GitHub repo is also available to increase transparency.

The logo png is renamed from sage-logo to scrubdash-logo to be more
descriptive. It was originally named sage-logo because the plant in the
logo looks relatively similar to a San Diego Sagewort, thus it was
technically a sage-logo.

The CSS styles file is reformatted to put colors at the bottom. This
allows future developers to be able to override the color of any text
since the color classes are at the bottom of the file, which gives the
color classes the highest precedent in stying.